predict/module/requirements.txt
requests
google-cloud-bigquery
pandas
xgboost
numpy
scikit-learn
sklearn
